pyproject.toml
[project]
name = "autoresearch"
version = "0.1.0"
description = "Autonomous pretraining research swarm"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kernels>=0.11.7",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "Pillow>=10.0.0",
    "pyarrow>=21.0.0",
    "requests>=2.32.0",
    "rustbpe>=0.1.0",
    "scikit-learn>=1.4.0",
    "tiktoken>=0.11.0",
    "timm>=1.0.0",
    "torch==2.9.1",
    "torchvision>=0.22.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128" },
]
torchvision = [
    { index = "pytorch-cu128" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
